+++ django/applications/catmaid/static/js/WindowMaker.js
@@ -118,21 +118,45 @@ var WindowMaker = new function()
     return {window: win, widget: instance};
   };
 
-  var createSelect = function(id, items, use_numbers) {
+  var createSelect = function(id, items, selectedValue) {
     var select = document.createElement('select');
     select.setAttribute("id", id);
     items.forEach(function(item, i) {
       var option = document.createElement("option");
-      option.text = item;
-      option.value = use_numbers ? i : item;
+      var itemType = typeof item;
+      var text, value;
+      if ('object' === itemType) {
+        text = item.title;
+        value = item.value;
+      } else {
+        text = item;
+        value = item;
+      }
+      option.text = text;
+      option.value = value;
+      if (option.value === selectedValue) {
+        option.defaultSelected = true;
+        option.selected = true;
+      }
       select.appendChild(option);
     });
     return select;
   };
 
-  var appendSelect = function(div, name, entries) {
-    var select = createSelect(div.id + "_" + name, entries, true);
+  var appendSelect = function(div, name, entries, title, value, onChangeFn) {
+    var select = createSelect(div.id + "_" + name, entries, value);
     div.appendChild(select);
+    if (title) {
+      select.title = title;
+    }
+    if (onChangeFn) {
+      select.onchange= onChangeFn;
+    }
+    var label = document.createElement('label');
+    label.setAttribute('title', title);
+    label.appendChild(document.createTextNode(name));
+    label.appendChild(select);
+    div.appendChild(label);
     return select;
   };
 
@@ -233,6 +257,8 @@ var WindowMaker = new function()
             return appendCheckbox(tab, e.label, e.title, e.value, e.onclickFn, e.left);
           case 'numeric':
             return appendNumericField(tab, e.label, e.title, e.value, e.postlabel, e.onchangeFn, e.length);
+          case 'select':
+            return appendSelect(tab, e.label, e.entries, e.title, e.value, e.onchangeFn);
           default: return undefined;
         }
       }
@@ -1347,15 +1373,38 @@ var WindowMaker = new function()
             WA.options.animation_back_forth = this.checked;
           }, false],
           {
-            type: 'checkbox',
-            label: 'Stepwise neuron visibility',
-            title: 'Start with one neuron visible, then make another neuron ' +
-                   'visible with each complete rotation.',
+            type: 'select',
+            label: 'Neuron visibility:',
+            entries: [
+              {title: 'Show all immeditely', value: 'all'},
+              {title: 'One per rotation', value: 'one-per-rotation'},
+              {title: 'N per rotation', value: 'n-per-rotation'}
+            ],
+            title: 'Select a neuron visibility pattern that is applied ' +
+                   'over the course of the animation.',
             value: o.animation_stepwise_visibility,
-            onclickFn: function() {
-              WA.options.animation_stepwise_visibility = this.checked;
-            },
-            left: false
+            onchangeFn: function(e) {
+              var type = this.value;
+              var options = {};
+              if ('one-per-rotation' === type) {
+                type = 'n-per-rotation';
+                options.n = 1;
+              } else if ('n-per-rotation' === type) {
+                // Ask for n
+                var dialog = new CATMAID.OptionsDialog();
+                dialog.appendMessage('Please enter the number of skeletons ' +
+                    'to make visible after one rotation.');
+                var nSkeletonsPerRot = dialog.appendField('Show n skeletons per rotation ',
+                   'show-n-skeletons-per-rot-' + WA.widgetID, 1, true);
+                dialog.onOK = function() {
+                  options.n = Number(nSkeletonsPerRot.value);
+                  WA.setAnimationNeuronVisibility(type, options);
+                };
+                dialog.show('auto', 'auto', true);
+                return;
+              }
+              WA.setAnimationNeuronVisibility(type, options);
+            }
           }
         ]);