+++ django/applications/catmaid/static/js/widgets/3dviewer.js
@@ -709,7 +709,8 @@
     this.animation_rotation_axis = "up";
     this.animation_rotation_speed = 0.01;
     this.animation_back_forth = false;
-    this.animation_stepwise_visibility = false;
+    this.animation_stepwise_visibility_type = 'all';
+    this.animation_stepwise_visibility_options = null;
     this.strahler_cut = 2; // to approximate twigs
   };
 
@@ -1138,6 +1139,35 @@
     }
 
     this.space.render();
+  };
+
+  /**
+   * Set neuron visibility with the help of a dialog in case the visibility type
+   * needs
+   */
+  WebGLApplication.prototype.setAnimationNeuronVisibility = function(type, options) {
+    var visibility = CATMAID.WebGLApplication.AnimationNeuronVisibilities[type];
+    if (!visibility) {
+      throw new CATMAID.ValueError("Unknown neuron visibility type: " + type);
+    }
+
+    // Bind visibility to widget and option object
+    this.options.animation_stepwise_visibility_type = type;
+    this.options.animation_stepwise_visibility_options = options;
+  };
+
+  WebGLApplication.AnimationNeuronVisibilities = {
+    'all': function() {},
+    'n-per-rotation': function(options, skeletonIds, visibleSkeletons, r) {
+      // Expect r to be the numnber of rotations done
+      var skeletonIndex = parseInt(r * options.n);
+      // Make next skeleton visible, if available
+      for (var i=0; i<options.n; ++i) {
+        if ((skeletonIndex + i) < skeletonIds.length) {
+          visibleSkeletons.push(skeletonIds[skeletonIndex + i]);
+        }
+      }
+    }
   };
 
   WebGLApplication.prototype.set_shading_method = function() {
@@ -4599,10 +4629,13 @@
 
     // Add a notification handler for stepwise visibility, if enabled and at least
     // one skeleton is loaded.
-    if (this.options.animation_stepwise_visibility) {
+    var visType = this.options.animation_stepwise_visibility_type;
+    if (visType !== 'all') {
       // Get current visibility map and create notify handler
       var visMap = this.space.getVisibilityMap();
-      options['notify'] = this.createStepwiseVisibilityHandler(visMap);
+      var visOpts = this.options.animation_stepwise_visibility_options;
+      options['notify'] = this.createStepwiseVisibilityHandler(visMap,
+          visType, visOpts);
       // Create a stop handler that resets visibility to the state we found before
       // the animation.
       options['stop'] = this.createVisibibilityResetHandler(visMap);
@@ -4615,8 +4648,11 @@
    * Create a notification handler to be used with animations that will make
    * an additional neuron visibile with every call.
    */
-  WebGLApplication.prototype.createStepwiseVisibilityHandler = function(visMap)
-  {
+  WebGLApplication.prototype.createStepwiseVisibilityHandler = function(
+      visMap, type, options) {
+    type = type || 'all';
+    options = options || {};
+
     // Get all visible skeletons
     var skeletonIds = Object.keys(this.space.content.skeletons)
         .filter(function(skid) {
@@ -4628,22 +4664,23 @@
       return function() {};
     }
 
-    // Only make first skeleton visible
-    var visibleSkeletons = [skeletonIds[0]];
-    this.space.setSkeletonVisibility(visMap, visibleSkeletons);
-
+    // Create function to animate neuron visibility
+    var visibility = CATMAID.WebGLApplication.AnimationNeuronVisibilities[type];
     var widget = this;
-
-    // Create function to make one skeleton visible per rotation
-    return function (r) {
-      // Expect r to be the numnber of rotations done
-      var skeletonIndex = parseInt(r);
-      // Make next skeleton visible, if available
-      if (skeletonIndex < skeletonIds.length) {
-        visibleSkeletons.push(skeletonIds[skeletonIndex]);
+    if ('all' === type) {
+      // If all should be shown, make all visible and return no-op
+      widget.space.setSkeletonVisibility(visMap, this.space.content.skeletons);
+      return function() {};
+    } else if (CATMAID.tools.isFn(visibility)) {
+      var visibleSkeletons = [];
+      return function (r) {
+        visibility(options, skeletonIds, visibleSkeletons, r);
         widget.space.setSkeletonVisibility(visMap, visibleSkeletons);
-      }
-    };
+      };
+    } else {
+      throw new CATMAID.ValueError('Don\'t recognize neuron visibility ' +
+          'animation mode "' + visibility + '".');
+    }
   };
 
   /**
@@ -4735,7 +4772,10 @@
           // Add a notification handler for stepwise visibility, if enabled and at least
           // one skeleton is loaded.
           if (stepVisibilityField.checked) {
-            options['notify'] = this.createStepwiseVisibilityHandler(visMap);
+            var visType = 'n-per-rotation';
+            var visOpts = {n: 1};
+            options['notify'] = this.createStepwiseVisibilityHandler(visMap,
+                visType, visOpts);
             // Create a stop handler that resets visibility to the state we found before
             // the animation.
             options['stop'] = this.createVisibibilityResetHandler(visMap);
@@ -5158,8 +5198,8 @@
   CATMAID.AnimationFactory.AxisRotation = function(camera, targetPosition, axis, rSpeed,
       backAndForth, notify)
   {
-    // Counts the number of rotations done
-    var numRotations = 0;
+    // Counts the number of rotations done after initialization
+    var numRotations = null;
 
     var targetDistance = camera.position.distanceTo(targetPosition);
     rSpeed = rSpeed || 0.01;